the function prints the elements in the nodes and return the number of nodes present.

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -12,7 +12,7 @@
 size_t print_listint(const listint_t *h)
 {
 	/*const listint_t *movptr = NULL;*/
-	int num_of_nodes = 0;
+	size_t num_of_nodes = 0;
 
 	/*if (h == NULL)
 		return (num_of_nodes);
